+++ commonjs/documentation/changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 <pre>
 v3.2.1  Added GeoFences APIs(create, update, query, remove) [TIMODOPEN-349]
+        Added PushNotifications APIs(setBadge, resetBadge) [TIMODOPEN-354]
 
 v3.2.0  Bumping version to match the TiSDK version this module is being shipped with